+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@vercel/blob": "^0.27.0",
     "embla-carousel-react": "^8.5.1",
     "next": "15.1.2",
     "react": "^19.0.0",
